Move climate API into separate file
