+++ src/openrct2/scripting/ScriptEngine.cpp
@@ -24,6 +24,7 @@
 #    include "../platform/Platform2.h"
 #    include "Duktape.hpp"
 #    include "ScCheats.hpp"
+#    include "ScClimate.hpp"
 #    include "ScConsole.hpp"
 #    include "ScContext.hpp"
 #    include "ScDate.hpp"
@@ -377,6 +378,7 @@ void ScriptEngine::Initialise()
 {
     auto ctx = static_cast<duk_context*>(_context);
     ScCheats::Register(ctx);
+    ScClimate::Register(ctx);
     ScConfiguration::Register(ctx);
     ScConsole::Register(ctx);
     ScContext::Register(ctx);
@@ -409,6 +411,7 @@ void ScriptEngine::Initialise()
     ScStaff::Register(ctx);
 
     dukglue_register_global(ctx, std::make_shared<ScCheats>(), "cheats");
+    dukglue_register_global(ctx, std::make_shared<ScClimate>(), "climate");
     dukglue_register_global(ctx, std::make_shared<ScConsole>(_console), "console");
     dukglue_register_global(ctx, std::make_shared<ScContext>(_execInfo, _hookEngine), "context");
     dukglue_register_global(ctx, std::make_shared<ScDate>(), "date");
